move content out of the source, and images
