+++ content/home/general/locationCost.md
