+++ content/home/persons/vermanen-debate.md
