+++ gatsby-config.js
@@ -7,13 +7,20 @@ module.exports = {
               resolve: `gatsby-source-filesystem`,
               options: {
                 name: `src`,
-                path: `${__dirname}/src/`,
+                path: `${__dirname}/content/`,
+              }
+            },
+            {
+              resolve: `gatsby-source-filesystem`,
+              options: {
+                name: `src`,
+                path: `${__dirname}/images/`,
               }
             },
             {
               resolve: `gatsby-plugin-favicon`,
               options: {
-                logo: "./src/images/favicon.png",
+                logo: "./images/favicon.png",
                 injectHTML: true,
                 icons: {
                   android: false,
@@ -37,7 +44,7 @@ module.exports = {
                 background_color: "#d7bda5",
                 theme_color: "#5e358c",
                 display: "minimal-ui",
-                icon: "src/images/favicon.png", // This path is relative to the root of the site.
+                icon: "images/favicon.png", // This path is relative to the root of the site.
               }
             },
             `gatsby-plugin-offline`, // After manifest
